[CWS] move common resolver func to common file
